finance_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.1.1
python-dotenv>=1.0
django-cors-headers>=4.0
firebase-admin>=6.5
